+++ test/gravitypull.sol
@@ -21,11 +21,10 @@ contract GravityFeed is Test, MyUtils {
         DAO.mintMembershipToken(Agent1);
         skip(1);
         DAO.signalInflation(10);
-        assertTrue(DAO.baseInflationRate() == 10, 'base inflation not set');
+        assertTrue(DAO.baseInflationRate() == 10, "base inflation not set");
 
         vm.stopPrank();
     }
-
 
     function _initNestedConstantRates(
         uint256 howMany,
@@ -67,7 +66,7 @@ contract GravityFeed is Test, MyUtils {
             IDAO20(DAOS[i].internalTokenAddress()).wrapMint(amtToTrickle);
             DAOS[i].signalInflation(inflationRate);
             console.log(i);
-            if(i < howMany-1) DAOS[i].distributiveSignal(distributionAmts);
+            if (i < howMany - 1) DAOS[i].distributiveSignal(distributionAmts);
             sum += DAOS[i].baseInflationRate();
             if (skipBetweenSignals > 0) skip(skipBetweenSignals);
         }
@@ -77,36 +76,36 @@ contract GravityFeed is Test, MyUtils {
         vm.stopPrank();
     }
 
-
-    function _DiferentiatedBalances(uint hM_) public returns (iInstanceDAO[] memory DAOS){
-
+    function _DiferentiatedBalances(uint256 hM_) public returns (iInstanceDAO[] memory DAOS) {
         uint256 timeStart = block.timestamp;
 
         DAOS = _initNestedConstantRates(hM_, 10, 20, 100 ether, 10, 10 days);
-    
+
         assertTrue(DAOS.length == hM_);
         uint256 timeNow = block.timestamp;
         assertTrue(timeNow == timeStart + (hM_ * 10 days));
 
-        uint i = DAOS.length;
+        uint256 i = DAOS.length;
 
-        for (i; i>=2;) {
-            iInstanceDAO D =  DAOS[i-1];
+        for (i; i >= 2;) {
+            iInstanceDAO D = DAOS[i - 1];
             if (D.parentDAO() == address(0)) break;
-            if (i > 2) assertTrue(D.parentDAO() == address(DAOS[i-2]));
-            
-            console.log(address(DAOS[i-1]));
+            if (i > 2) assertTrue(D.parentDAO() == address(DAOS[i - 2]));
+
+            console.log(address(DAOS[i - 1]));
 
             address[] memory path = new address[](hM_);
-            path = O.getTrickleDownPath(address(DAOS[i-1]));
+            path = O.getTrickleDownPath(address(DAOS[i - 1]));
 
-            unchecked { --i;}
+            unchecked {
+                --i;
+            }
         }
-
     }
 
     /// @dev fuzz : uint howMany
-    uint howMany =12;
+    uint256 howMany = 12;
+
     function testDiferentiatedBalances() public {
         // vm.assume(howMany < 99);
         // vm.assume(howMany > 2);
@@ -114,24 +113,23 @@ contract GravityFeed is Test, MyUtils {
         DsszNuts = _DiferentiatedBalances(howMany);
 
         address[] memory fullPath = new address[](DsszNuts.length);
-        iInstanceDAO lastNut = DsszNuts[DsszNuts.length-1];
+        iInstanceDAO lastNut = DsszNuts[DsszNuts.length - 1];
         fullPath = O.getTrickleDownPath(address(lastNut));
 
-        assertTrue(fullPath[fullPath.length-1] == address(0), 'chain end is not 0');
-        assertTrue(fullPath[0] == lastNut.parentDAO(), 'fist in path is not parent');
+        assertTrue(fullPath[fullPath.length - 1] == address(0), "chain end is not 0");
+        assertTrue(fullPath[0] == lastNut.parentDAO(), "fist in path is not parent");
 
-
-        uint256 i=0;
-        for (i; i < fullPath.length-2; ) {
-            IERC20 baseT = IERC20(iInstanceDAO(fullPath[i+1]).baseTokenAddress());
-            IERC20 internalT = IERC20(iInstanceDAO(fullPath[i+1]).internalTokenAddress());
-            uint baseBalanceBeforeFeed = baseT.balanceOf(fullPath[i+1]);
-            uint internalTBeforeFeed =  internalT.balanceOf(fullPath[i+1]);
-            assertTrue(fullPath[i+1] == iInstanceDAO(fullPath[i]).parentDAO(), "expected parent - subdao relations");
-            iInstanceDAO(fullPath[i+1]).feedMe();
+        uint256 i = 0;
+        for (i; i < fullPath.length - 2;) {
+            IERC20 baseT = IERC20(iInstanceDAO(fullPath[i + 1]).baseTokenAddress());
+            IERC20 internalT = IERC20(iInstanceDAO(fullPath[i + 1]).internalTokenAddress());
+            uint256 baseBalanceBeforeFeed = baseT.balanceOf(fullPath[i + 1]);
+            uint256 internalTBeforeFeed = internalT.balanceOf(fullPath[i + 1]);
+            assertTrue(fullPath[i + 1] == iInstanceDAO(fullPath[i]).parentDAO(), "expected parent - subdao relations");
+            iInstanceDAO(fullPath[i + 1]).feedMe();
             skip(10);
-            uint baseBalanceAfterFeed = baseT.balanceOf(fullPath[i+1]);
-            uint internalTAfterFeed =  internalT.balanceOf(fullPath[i+1]);
+            uint256 baseBalanceAfterFeed = baseT.balanceOf(fullPath[i + 1]);
+            uint256 internalTAfterFeed = internalT.balanceOf(fullPath[i + 1]);
             console.log(baseBalanceBeforeFeed);
             console.log(internalTBeforeFeed);
             console.log("------beforeFeed^----afterFeed >");
@@ -141,13 +139,9 @@ contract GravityFeed is Test, MyUtils {
             assertTrue(baseBalanceAfterFeed > baseBalanceBeforeFeed);
             assertTrue(internalTBeforeFeed == internalTAfterFeed);
 
-            unchecked { ++ i;}
-
+            unchecked {
+                ++i;
+            }
         }
-
     }
-
-
 }
-
-
